add log on systemcontrol & add permission lost

+++ frameworks/common/src/avmeta_data.cpp
@@ -132,7 +132,7 @@ bool AVMetaData::UnmarshallingExceptImg(MessageParcel& data, AVMetaData& metaOut
         return false;
     }
     SLOGI("get mask with %{public}s", mask.c_str());
-    for (size_t i = 0; i < maskSize; ++i) {
+    for (int32_t i = 0; i < maskSize; ++i) {
         if (mask[i] == '1') {
             metaOut.metaMask_.flip(maskSize - i - 1);
         }
